WIP: Refactor OpenPositions component
