--- src/components/CareersPage/BenefitsData.tsx
@@ -1,27 +1,0 @@
-export const benefitsData = [
-  {
-    benefit: 'Flexible work hours',
-    id: 1,
-    text: 'We believe that your job should make the "life" portion of your work-life balance a little more enjoyable and a little less stressful. This means you shouldn not have to retrofit your life to your work, rather, the other way around, which is why we offer a flexible work schedule and why our employees choose the Base Two location that best suits them.',
-  },
-  {
-    benefit: 'Free insurance',
-    id: 2,
-    text: 'Health insurance is important, but nobody likes losing a portion of their paycheck to it. Base Two pays 100% of the premiums for employee health, vision, and dental insurance so that you dont need to worry about it.',
-  },
-  {
-    benefit: '401K plan',
-    id: 3,
-    text: 'Saving for retirement is important, and Base Twos 401k plan provides a method to save with pre-tax money. To help with your savings Base Two contributes 3% of your salary to your Base Two 401k plan.',
-  },
-  {
-    benefit: 'Generous time off',
-    id: 4,
-    text: "Employees are encouraged to take an average of 3 weeks of vacation a year. We also close the office for 12 total holidays and offer unlimited sick days. Let's face it, we all get sick. ",
-  },
-  {
-    benefit: 'Choice of equipment',
-    id: 5,
-    text: 'Base Two does not have standard-issue equipment to hand new employees, but rather allows each employee to select the equipment that will work best for them. If it makes you happy and helps you do your work, then it’s worth it.',
-  },
-];
