+++ src/components/CareersPage/OpenPositions.tsx
@@ -1,11 +1,12 @@
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
-import type { FC } from 'react';
+import { FC } from 'react';
 
 import { Card, Heading, Text } from '~/components';
 import { atMinTablet, atMinXL, colors, spacing } from '~/theme';
 
 import { Button } from '../ContactPage/ContactForm';
+import { PositionsData } from './CareersData';
 
 export const PositionContainer = styled.div`
   background: ${colors.darkBlueAlt};
@@ -13,50 +14,13 @@ export const PositionContainer = styled.div`
   padding: ${spacing.md} ${spacing.sm};
   width: 100%;
   ${atMinTablet} {
-    padding: ${spacing.xxl1} 12rem;
+    padding: ${spacing.xxl} ${spacing.xxl3};
   }
 `;
 
-export type PositionProps = {
-  role: string;
-  description: string;
-};
+export const OpenPositions: FC = () => {
+  const jobs = PositionsData;
 
-export const Position: FC<PositionProps> = ({ description, role }) => {
-  return (
-    <PositionContainer>
-      <Heading
-        as="h3"
-        color="coral"
-        css={css`
-          font-variation-settings: 'wght' 700;
-        `}
-        variant="h3"
-      >
-        {role}
-      </Heading>
-      <Text
-        css={css`
-          margin-bottom: ${spacing.lg};
-        `}
-      >
-        {description}
-      </Text>
-      <Button
-        css={css`
-          width: 100%;
-          ${atMinTablet} {
-            width: auto;
-          }
-        `}
-      >
-        Apply
-      </Button>
-    </PositionContainer>
-  );
-};
-
-export const OpenPositions: FC = ({ children }) => {
   return (
     <section
       css={css`
@@ -87,7 +51,34 @@ export const OpenPositions: FC = ({ children }) => {
           besides and beyond development do we need the most.
         </Heading>
       </Card>
-      {children}
+
+      {jobs.map((job) => {
+        return (
+          <>
+            <PositionContainer key={job.id}>
+              <Heading
+                as="h3"
+                color="coral"
+                css={css`
+                  font-variation-settings: 'wght' 700;
+                `}
+                variant="h3"
+              >
+                {job.job}
+              </Heading>
+
+              <Text
+                css={css`
+                  margin-bottom: ${spacing.lg};
+                `}
+              >
+                {job.desc}
+              </Text>
+              <Button>Apply</Button>
+            </PositionContainer>
+          </>
+        );
+      })}
     </section>
   );
 };
