+++ src/components/CareersPage/index.tsx
@@ -6,7 +6,7 @@ import { Layout } from '~/components';
 import { PageHero } from '../PageHero';
 import { Benefits } from './Benefits';
 import { Benefit, BenefitsAccordion } from './BenefitsAccordion';
-import { OpenPositions, Position } from './OpenPositions';
+import { OpenPositions } from './OpenPositions';
 import { WorkingAtB2 } from './WorkingAtB2';
 
 export const CareersPage: FC = () => {
@@ -60,24 +60,7 @@ export const CareersPage: FC = () => {
         </BenefitsAccordion>
       )}
 
-      <OpenPositions>
-        <Position
-          description="We are always looking for amazing people to work with us in Columbus, OH; Pittsburgh, PA; and Chicago, IL. If you are a talented and creative individual with strong people skills and experience building web applications with the latest front-end technologies, please apply."
-          role="Senior Software Engineer"
-        />
-        <Position
-          description="We are always looking for amazing people to work with us in Columbus, OH; Pittsburgh, PA; and Chicago, IL. If you are a talented and creative individual with strong people skills and experience building web applications with the latest front-end technologies, please apply."
-          role="Software Engineer"
-        />
-        <Position
-          description="We are always looking for amazing people to work with us in Columbus, OH; Pittsburgh, PA; and Chicago, IL. If you are a talented and creative individual with strong people skills and experience building web applications with the latest front-end technologies, please apply."
-          role="Jr. Software Engineer"
-        />
-        <Position
-          description="We are always looking for amazing people to work with us in Columbus, OH; Pittsburgh, PA; and Chicago, IL. If you are a talented and creative individual with strong people skills and experience building web applications with the latest front-end technologies, please apply."
-          role="Mid/Senior Software Engineer"
-        />
-      </OpenPositions>
+      <OpenPositions />
     </Layout>
   );
 };
